+++ ft_memset.c
@@ -6,16 +6,16 @@
 /*   By: omadali <adalomer60@gmail.com>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/27 14:22:09 by omadali           #+#    #+#             */
-/*   Updated: 2024/10/03 21:20:39 by omadali          ###   ########.fr       */
+/*   Updated: 2024/10/04 02:02:24 by omadali          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memset(void *s, int c, int n)
+void	*ft_memset(void *s, int c, size_t n)
 {
-	int			a;
-	char		*b;
+	size_t	a;
+	char	*b;
 
 	if (!s)
 		return (0);
